projs/mlflow-appinsights/pyproject.toml
[project]
name = "mlflow-appinsights"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-core==1.38.0",
    "azure-mgmt-web==8.0.0",
    "azure-monitor-opentelemetry==1.4.1",
    "azure-storage-file-datalake==12.17.0",
    "cloudpickle==3.0.0",
    "cryptography==43.0.3",
    "mlflow==3.9.0",
    "pandas==2.2.3",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ruff>=0.15.0",
]
